+++ sclc/include/ds/dynamic_array.h
@@ -1,28 +1,28 @@
 #ifndef DYNAMIC_ARRAY
 #define DYNAMIC_ARRAY
 
-#include <stddef.h>
+#include "common.h"
 
 typedef struct dynamic_array {
   void *items;
-  size_t item_size;
-  size_t count;
-  size_t capacity;
+  u64 item_size;
+  u64 count;
+  u64 capacity;
 } dynamic_array;
 
-void dynamic_array_init(dynamic_array *da, size_t size);
+void dynamic_array_init(dynamic_array *da, u64 size);
 
-int dynamic_array_get(dynamic_array *da, size_t index, void *item);
+u32 dynamic_array_get(dynamic_array *da, u64 index, void *item);
 
-int dynamic_array_set(dynamic_array *da, size_t index, void *item);
+u32 dynamic_array_set(dynamic_array *da, u64 index, void *item);
 
-int dynamic_array_append(dynamic_array *da, void *item);
+u32 dynamic_array_append(dynamic_array *da, void *item);
 
-int dynamic_array_insert(dynamic_array *da, size_t index, void *item);
+u32 dynamic_array_insert(dynamic_array *da, u64 index, void *item);
 
-int dynamic_array_remove(dynamic_array *da, size_t index);
+u32 dynamic_array_remove(dynamic_array *da, u64 index);
 
-int dynamic_array_pop(dynamic_array *da, void *item);
+u32 dynamic_array_pop(dynamic_array *da, void *item);
 
 void dynamic_array_free(dynamic_array *da);
