+++ sclc/include/lexer.h
@@ -5,9 +5,8 @@
 #ifndef LEXER_H
 #define LEXER_H
 
+#include "common.h"
 #include "ds/dynamic_array.h"
-
-#include <stddef.h>
 
 /*
  * @struct lexer: maintains state of the lexer for tokenizing the source buffer.
@@ -17,15 +16,15 @@ typedef struct lexer {
    * Pointer to source buffer and its size in bytes.
    */
   const char *buffer;
-  size_t buffer_len;
+  u64 buffer_len;
 
   /*
    * Data concerned with current lexer state.
    */
-  size_t line;     // <-- tracks the current line
-  size_t pos;      // <-- current position in buffer
-  size_t read_pos; // <-- next read position (usually pos + 1)
-  char ch;         // <-- character at buffer[read_pos]
+  u64 line;     // <-- tracks the current line
+  u64 pos;      // <-- current position in buffer
+  u64 read_pos; // <-- next read position (usually pos + 1)
+  char ch;      // <-- character at buffer[read_pos]
 } lexer;
 
 /*
@@ -35,7 +34,7 @@ typedef struct lexer {
  * @param buffer_len size of buffer (in bytes).
  * @param tokens: dynamic_array of tokens (should be initialized).
  */
-void lexer_tokenize(const char *buffer, size_t buffer_len,
-                    dynamic_array *tokens, const char *include_dir);
+void lexer_tokenize(const char *buffer, u64 buffer_len, dynamic_array *tokens,
+                    const char *include_dir);
 
 #endif // !LEXER_H
